requirements.txt
openai>=1.100.0
python-dotenv>=1.0.1
pandas>=2.2.2
numpy>=1.26.4
pysd>=3.14.2
openpyxl>=3.1.5
matplotlib>=3.9.0
pydantic>=2.8.2
streamlit>=1.38.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
